changing in cart

+++ cart.js
@@ -1,5 +1,6 @@
 function back() {
     window.location = "index.html"
+    localStorage.setItem('addItem', JSON.stringify(cartArr));
 
 
 }
@@ -11,21 +12,37 @@ console.log(cartArr);
 
 
 const div = document.querySelector('#cart')
+const totalAmounts = document.querySelector('.total-amount');
+
 
 function renderCart() {
-    for (let i = 0; i < cartArr.length; i++) {
-        div.innerHTML += `<div class="container ">
-       
-        <img class="images" src="${cartArr[i].img}">
-        <h3 >${cartArr[i].brand} ${cartArr[i].model}</h3>
-       
-        <h3 class="price">Price: ${cartArr[i].price} </h3>
-        <h3 class="quantity">Quantity: ${cartArr[i].quantity}</h3>
-        <h3 class="price">Total Price: ${cartArr[i].price * cartArr[i].quantity}  </h3>
-        <button class="bg-sky-400 hover:bg-sky-600 text-white font-bold py-2 px-4 rounded " onclick="deleteItem(${i})">Delete</button>
-        <button class="bg-sky-400 hover:bg-sky-600 text-white font-bold py-2 px-4 rounded " onclick="increaseQuantity(${i})">+</button>
-        <button class="bg-sky-400 hover:bg-sky-600 text-white font-bold py-2 px-4 rounded " onclick="decreaseQuantity(${i})">-</button>
-        </div>`
+    let totalAmount = 0;
+    totalAmount.innerHTML = '';
+    if (cartArr.length > 0) {
+        for (let i = 0; i < cartArr.length; i++) {
+            totalAmount += cartArr[i].price * cartArr[i].quantity
+            div.innerHTML += `<div class="container ">
+            <img class="images" src="${cartArr[i].img}">
+            <h3 class="brand" >${cartArr[i].brand} ${cartArr[i].model}</h3>
+            <h3 class="price"><s>Price: ${cartArr[i].price}</s> </h3>
+            <h3 class="price">Total Price: ${cartArr[i].price * cartArr[i].quantity}  </h3>
+            <div class="quantity-div">
+            <h3 class="quantity">Quantity : ${cartArr[i].quantity}</h3>
+           <div class="">
+           <button class="text-xl bg-sky-400 hover:bg-sky-600 text-white font-bold  rounded decrease-btn " onclick="increaseQuantity(${i})">+</button>
+           <button class="text-xl bg-sky-400 hover:bg-sky-600 text-white font-bold  rounded decrease-btn "onclick="decreaseQuantity(${i})">-</button>
+           </div>
+            </div>
+         
+            <div class="cart-div">
+            <button class="bg-sky-400 hover:bg-sky-600 text-white font-bold py-2 px-4 rounded " onclick="deleteItem(${i})">Delete</button>
+            </div>
+          
+            </div>`
+            totalAmounts.innerHTML = `<span>Total Amount</span> : ${totalAmount}`
+        }
+    } else {
+        div.innerHTML = `  <h2>No item found</h2>`
 
     }
 
@@ -64,3 +81,7 @@ function decreaseQuantity(index) {
     }
 
 }
+
+window.onbeforeunload = function () {
+    localStorage.setItem('addItem', JSON.stringify(cartArr));
+};
